+++ src/modules/boost/index.ts
@@ -2,10 +2,12 @@ import BoostComposerStack from './boost-composer/navigator';
 import BoostConsoleScreen from './boost-console/screens/BoostConsoleScreen';
 import SingleBoostConsoleScreen from './boost-console/screens/SingleBoostConsoleScreen';
 import boostedContentService from './services/boosted-content.service';
+import { cleanBoosts } from './utils/clean-boosts';
 
 export {
   BoostComposerStack,
   BoostConsoleScreen,
   SingleBoostConsoleScreen,
   boostedContentService,
+  cleanBoosts,
 };