+++ src/modules/boost/services/boosted-content.service.ts
@@ -1,9 +1,8 @@
-import { Platform } from 'react-native';
-import apiService from '~/common/services/api.service';
-import blockListService from '~/common/services/block-list.service';
+import FeedsService from '~/common/services/feeds.service';
+import logService from '~/common/services/log.service';
 import sessionService from '~/common/services/session.service';
-import { storages } from '~/common/services/storage/storages.service';
-import ActivityModel from '~/newsfeed/ActivityModel';
+import type ActivityModel from '~/newsfeed/ActivityModel';
+import { cleanBoosts } from '../utils/clean-boosts';
 
 /**
  * Boosted content service
@@ -14,6 +13,12 @@ class BoostedContentService {
    * @var {number}
    */
   offset: number = -1;
+
+  /**
+   * Feed service
+   * @var {FeedsService}
+   */
+  feedsService?: FeedsService;
 
   /**
    * Boosts
@@ -27,30 +32,49 @@ class BoostedContentService {
   updating = false;
 
   /**
-   * Initialize if necessary
+   * Reload boosts list
    */
-  init() {
+  load = async (): Promise<any> => {
+    this.init();
     if (!sessionService.userLoggedIn || sessionService.switchingAccount) {
       return;
     }
-    this.loadCached();
-    return this.update();
+    try {
+      const done = await this.feedsService!.setOffset(0).fetchLocal();
+
+      if (!done) {
+        await this.update();
+      } else {
+        this.boosts = cleanBoosts(await this.feedsService!.getEntities());
+        this.update();
+      }
+    } catch (err) {
+      logService.exception('[BoostedContentService]', err);
+    }
+  };
+
+  /**
+   * Initialize if necessary
+   */
+  init() {
+    if (!this.feedsService) {
+      this.feedsService = new FeedsService();
+      this.feedsService
+        .setLimit(24)
+        .setOffset(0)
+        .setPaginated(false)
+        .setEndpoint('api/v3/boosts/feed')
+        .setDataProperty('boosts')
+        .setParams({ location: 1 });
+    }
   }
 
   /**
    * Clear the service
    */
   clear() {
-    this.boosts = [];
-  }
-
-  loadCached() {
-    const boosts = storages.session?.getArray<ActivityModel>(
-      'BoostServiceCache',
-    );
-    if (boosts) {
-      this.boosts = ActivityModel.createMany(boosts);
-    }
+    this.feedsService?.clear();
+    delete this.feedsService;
   }
 
   /**
@@ -59,17 +83,8 @@ class BoostedContentService {
   async update() {
     try {
       this.updating = true;
-      const response = await apiService.get<any>('api/v3/boosts/feed', {
-        location: 1,
-      });
-      if (response?.boosts) {
-        const filteredBoosts = cleanBoosts(response.boosts.map(b => b.entity));
-
-        this.boosts = ActivityModel.createMany(filteredBoosts);
-
-        // cache boosts
-        storages.session?.setArray('BoostServiceCache', filteredBoosts);
-      }
+      await this.feedsService!.fetch();
+      this.boosts = cleanBoosts(await this.feedsService!.getEntities());
     } finally {
       this.updating = false;
     }
@@ -92,21 +107,6 @@ class BoostedContentService {
 
     return this.boosts[this.offset];
   }
-}
-
-export function cleanBoosts(
-  boosts: Array<ActivityModel>,
-): Array<ActivityModel> {
-  return boosts.filter((entity: ActivityModel) => {
-    entity.boosted = true;
-    // remove NSFW on iOS
-    if (Platform.OS === 'ios' && entity.nsfw && entity.nsfw.length) {
-      return false;
-    }
-    return entity.type === 'user'
-      ? false
-      : !blockListService.has(entity.ownerObj?.guid);
-  });
 }
 
 export default new BoostedContentService();