+++ CHANGELOG.md
@@ -2,6 +2,17 @@
 
 Minds Mobile
 
+## 4.33.0 (2023-03-23)
+
+### Features ✨ (2 changes)
+
+- [Boost partner in-feed notice added](minds/mobile-native!2182)
+- [Added background color to Mind+ comments](minds/mobile-native!2170)
+
+### Changed 🔧 (1 change)
+
+- [Disable cash boosts and upgrade screen for play store](minds/mobile-native!2190)
+
 ## 4.32.5 (2023-03-20)
 
 ### Fixed 🐛 (2 changes)